pyproject.toml
[project]
name = "telegram-image-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Bot Framework
    "aiogram==3.22.0",
    "aiohttp==3.10.11",
    # Database
    "sqlalchemy==2.0.44",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    # Task Queue
    "celery==5.4.0",
    "redis==5.2.0",
    # Image Processing
    "Pillow==10.1.0",
    "rembg==2.0.64",
    "opencv-python-headless==4.10.0.84",
    "numpy==2.1.3",
    # Object Storage
    "minio==7.2.18",
    # Configuration & Validation
    "pydantic==2.8.2",
    "pydantic-settings==2.6.1",
    # Monitoring & Utilities
    "prometheus-client==0.21.1",
    "python-json-logger==2.0.7",
    "structlog==24.4.0",
    "psycopg2-binary>=2.9.11",
    "uvloop>=0.22.1",
    "loguru>=0.7.3",
    "tenacity>=9.1.2",
]
